Fix example reference in "Get the code of conduct for a repository" override
At the moment, our override for the "Get the code of conduct for a
repository" API is broken for many versions because it refers
to a `code-of-conduct-2` example which isn't in current versions
of our bundled schemas.

This swaps out that reference to refer to plain-ol'
`code-of-conduct`. I've done a quick comparison and the two
historic files seem to be functionally identical 🤷‍♀️