Fix the "Get the code of conduct for a repository" override (#195)
* Add "Get the code of conduct for a repository" to the generated schema for GHES 3.3

We have "overrides" which can add, remove or update parts of the
schema.

We currently have an override which adds the
"Get the code of conduct for a repository" operation to the
schemas for GitHub.com, GHAE and GHES 3.2. These aren't included
in the normal bundled schemas because they are marked as
deprecated (and thus not to be included in documentation), but we
still want to support them in Octokit for users who rely on them.

This starts using this override for GHES 3.3 as well, since the
"Get the code of conduct for a repository" API is included in that
version as well. It does not appear in GHES 3.4 or later.

* Fix example reference in "Get the code of conduct for a repository" override

At the moment, our override for the "Get the code of conduct for a
repository" API is broken for many versions because it refers
to a `code-of-conduct-2` example which isn't in current versions
of our bundled schemas.

This swaps out that reference to refer to plain-ol'
`code-of-conduct`. I've done a quick comparison and the two
historic files seem to be functionally identical 🤷‍♀️