build: refactor overrides to use shared helpers with helpful validations
This refactors our OpenAPI overrides to use shared helpers
(`rewriteOperationId`, `addOperation`) rather than bespoke code
for each and every override.

These helpers include simple validations - for example, if you
try to change an operation ID and the operation doesn't exist,
then it will error.

This aims to increase the amount of code that is shared and
enforce good hygiene - specifically, only having overrides that
are valid and actually get applied. If you don't want to apply
an override, then you can use an `if` condition rather than
having the code as a no-op.