build: force a human to check overrides are valid when there is a new GHES version
This updates the overrides script to explicitly state what GHES
versions it supports.

This means that, when there is a new GHES version, the build
process will fail until a human looks at the code and checks that
the provided overrides still make sense for the new version.

This avoids applying overrides that don't make sense or having
dead code hanging around for a long time.