Perform a dry-run of the update process in CI when changes are made to the build process
We have a regular update task which runs once a day, fetches
updates to https://github.com/github/rest-api-description and
creates a PR in this repo.

When we make changes to the build scripts, dependencies, etc.,
it is possible to break this regular task accidentally. The only
way you know it's broken is when the task runs a day later and
it explodes.

I made this mistake twice today with dependency ipdates: #216
and #215.

This starts performing a dry-run of the update task whenever
changes are pushed outside of `cache/` and `generated/`, doing
everything except pushing and creating a PR. This can help to
provide greater confidence and safety when updating the code.

Fixes #223.