Handle GHEC-only SCIM APIs which have `$ref` keys which are not OpenAPI references
The build script currently assumes that `$ref` keys that it finds
in the OpenAPI specification are OpenAPI references to be unfurled.
This isn't the case for some real attributes in the SCIM API called
`$ref` 😅

This updates the build script to only try to expand a reference
that looks like an OpenAPI reference (it is a string and begins
with `#/`).